Sprint 4 almost done

+++ src/modules/campaigns/services/DeleteCampaignsService.ts
@@ -2,13 +2,9 @@ import { inject, injectable } from 'tsyringe';
 
 import { Campaigns } from '@prisma/client';
 
-// import AppError from '@shared/errors/AppError';
+import AppError from '@shared/errors/AppError';
 
 import ICampaignsRepository from '../repositories/ICampaignsRepository';
-
-interface IRequest {
-  id: string;
-}
 
 @injectable()
 export default class DeleteCampaignsService {
@@ -17,9 +13,10 @@ export default class DeleteCampaignsService {
     private CampaignsRepository: ICampaignsRepository,
   ) { }
 
-  public async execute({
-    id,
-  }: IRequest): Promise<Campaigns> {
+  public async execute(id: string): Promise<Campaigns> {
+    const campaignExists = await this.CampaignsRepository.findByID(id);
+
+    if (!campaignExists) throw new AppError('A campaign with this id does not exist');
     const campaign = await this.CampaignsRepository.delete(id);
 
     return campaign;
